Placed the code for the password reset email into a suitable place- accounts app. Moved imports and url's from home app over to main app url.py to match the other links referring to the same and also the accounts views.py
