+++ accounts/views.py
@@ -6,6 +6,17 @@ from accounts.forms import UserLoginForm, UserRegistrationForm
 from userProfile.models import UserProfile, UserAddress
 from userProfile.urls import user_profile 
 
+# Used for the password reset
+from django.core.mail import send_mail, BadHeaderError
+from django.http import HttpResponse
+from django.contrib.auth.forms import PasswordResetForm
+from django.contrib.sites.shortcuts import get_current_site
+from django.template.loader import render_to_string
+from django.db.models.query_utils import Q
+from django.utils.http import urlsafe_base64_encode
+from django.contrib.auth.tokens import default_token_generator
+from django.utils.encoding import force_bytes
+
 
 # Views for the accounts app.
 
@@ -44,8 +55,6 @@ def login(request):
     else:
         login_form = UserLoginForm()
     return render(request, 'login.html', {"login_form" : login_form})
-
-
 
 
 def registration(request):
@@ -89,3 +98,36 @@ def makeProfile(request):
     uar = UserAddress()
     uar.person = UserProfile.objects.get(user=upr.user)
     uar.save()
+
+
+def password_reset_request(request):
+	if request.method == "POST":
+		password_reset_form = PasswordResetForm(request.POST)
+		
+		if password_reset_form.is_valid():
+			data = password_reset_form.cleaned_data['email']
+			associated_users = User.objects.filter(Q(email=data))
+			
+			if associated_users.exists():
+				for user in associated_users:
+					subject = "Password Reset Requested"
+					email_template_name = "registration/password_reset_email.html"
+					context = {
+					"email":user.email,
+					'domain':get_current_site(request),
+					'site_name': 'The Scent Queen',
+					"uid": urlsafe_base64_encode(force_bytes(user.pk)),
+					"user": user,
+					'token': default_token_generator.make_token(user),
+					'protocol': 'https',
+					}
+					email = render_to_string(email_template_name, context)
+					
+					try:
+						send_mail(subject, email, 'info.thescentqueen@gmail.com' , [user.email], fail_silently=False)
+					except BadHeaderError:
+						return HttpResponse('Invalid header found.')
+					return redirect ("/password_reset/done/")
+					
+	password_reset_form = PasswordResetForm()
+	return render(request=request, template_name="registration/password_reset_form.html", context={"password_reset_form":password_reset_form})
