+++ home/urls.py
@@ -4,5 +4,4 @@ from . import views
 urlpatterns = [
     path('', views.home, name='home'),
     path('about', views.about, name="about"),
-    path("password_reset", views.password_reset_request, name="password_reset")
     ]
