+++ templates/registration/password_reset_email.html
@@ -6,7 +6,9 @@ You're receiving this e-mail because you requested a password reset for your use
 
 Please go to the following page and choose a new password:
 {% block reset_link %}
+
 {{ protocol }}://{{ domain }}{% url 'password_reset_confirm' uidb64=uid token=token %}
+
 {% endblock %}
 
 Thanks for using our site!
