--- templates/registration/password_reset_email.txt
@@ -1,16 +1,0 @@
-{% autoescape off %}
-
-Hello {{ user.username }},
-
-You're receiving this e-mail because you requested a password reset for your user account at ScentQueen!
-
-Please go to the following page and choose a new password:
-{% block reset_link %}
-{{ protocol }}://{{ domain }}{% url 'password_reset_confirm' uidb64=uid token=token %}
-{% endblock %}
-
-Thanks for using our site!
-
-The ScentQueen.
-
-{% endautoescape %}
